+++ packages/ssx/src/__fixtures__/src/pages/index.js
@@ -1,10 +1,12 @@
+import "@inglorious/logo/style.css"
+
 import { html } from "@inglorious/web"
 
 import { nav } from "../components/nav.js"
 
 export const index = {
-  render() {
+  render(_, api) {
     return html`<h1>Index</h1>
-      ${nav.render()}`
+      <div>${api.render("logo")} ${nav.render()}</div>`
   },
 }