+++ packages/ssx/src/render/index.js
@@ -52,6 +52,6 @@ export async function renderPage(store, page, entity, options = {}) {
     styles,
     head,
     scripts,
-    ssxEntity: { [moduleName]: entity },
+    ssxPage: { [moduleName]: entity },
   })
 }