+++ examples/apps/web-components-ssx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-components-ssx
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/ssx@1.10.0
+
 ## 0.1.1
 
 ### Patch Changes